Add simple code generation

+++ interpreter/include/cpu.h
@@ -3,6 +3,6 @@
 #include <stdint.h>
 
 int cpu_execute(uint32_t* instructions, int instructions_size,
-        uint32_t* memory, int memory_size,
+        uint8_t* memory, int memory_size,
         int entry_point);
 #endif
